fix: getFileURL
