+++ packages/plugins/@nocobase/plugin-file-manager/src/server/storages/index.ts
@@ -7,8 +7,8 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-import Application from '@nocobase/server';
 import { StorageEngine } from 'multer';
+import urlJoin from 'url-join';
 
 export interface StorageModel {
   id?: number;
@@ -42,9 +42,13 @@ export abstract class StorageType {
 
   getFileData?(file: { [key: string]: any }): { [key: string]: any };
   getFileURL(file: AttachmentModel, preview?: boolean): string | Promise<string> {
-    return `${(this.storage.baseUrl || '').replace(/\/|$/, '')}/${
-      file.path ? `${encodeURI(file.path)}/` : ''
-    }${encodeURIComponent(file.filename)}${(preview && this.storage.options.thumbnailRule) || ''}`;
+    const keys = [
+      this.storage.baseUrl,
+      file.path && encodeURI(file.path),
+      encodeURIComponent(file.filename),
+      preview && this.storage.options.thumbnailRule,
+    ].filter(Boolean);
+    return urlJoin(keys);
   }
 }
